# api/openapi.yaml
openapi: 3.1.0
info:
  title: TX Carpool API
  version: 1.0.0
  description: |
    HTTP contract for the TX Carpool Telegram Mini App. User and admin endpoints
    use a Secure, HttpOnly session cookie established from validated Telegram
    Mini App `initData`. Provider callbacks use provider-specific signatures and
    are never authenticated by the browser session.

    All identifiers and timestamps are strings. TXB values use integer hundredths
    encoded as decimal strings. Traffic byte counts are decimal strings. The
    server is authoritative for prices, exchange rates, entitlement dates, and
    payment state.
  license:
    name: Proprietary
servers:
  - url: /
    description: Same-origin API
...
security:
  - cookieSession: []
paths:
  /api/v1/auth/telegram:
    post:
      tags: [Authentication]
      operationId: authenticateTelegram
      summary: Exchange Telegram Mini App init data for a session
      description: |
        Validates the raw query string according to Telegram's Mini App algorithm,
        rejects `auth_date` older than five minutes, upserts the Telegram identity,
        and sets a seven-day Secure, HttpOnly, SameSite session cookie. Data from
        `initDataUnsafe` is never accepted.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelegramAuthRequest'
      responses:
        '200':
          description: Authenticated; the session cookie is set.
          headers:
            Set-Cookie:
              description: Opaque session cookie. Its value is never exposed to JavaScript.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthState'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/me:
    get:
      tags: [Authentication]
      operationId: getMe
      summary: Get the current identity and onboarding state
      responses:
        '200':
          description: Current authenticated state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthState'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/onboarding/invites:
    post:
      tags: [Onboarding]
      operationId: createOnboardingInvites
      summary: Create identity-bound group and channel join-request links
      description: |
        Creates or replaces 30-minute join-request links for the authenticated
        Telegram identity. Only that identity may be approved. Links are revoked
        after approval or replacement.
      responses:
        '200':
          description: Active one-time join-request links.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingInvites'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          $ref: '#/components/responses/UpstreamFailure'
  /api/v1/onboarding/membership/check:
    post:
      tags: [Onboarding]
      operationId: checkOnboardingMembership
      summary: Recheck canonical group and channel membership
      description: Calls Telegram `getChatMember` for both configured chats and advances onboarding only when both memberships are valid.
      responses:
        '200':
          description: Refreshed canonical membership and user state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipState'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          $ref: '#/components/responses/UpstreamFailure'
  /api/v1/onboarding/username:
    put:
      tags: [Onboarding]
      operationId: reserveUsername
      summary: Reserve the immutable service username
      description: |
        Accepts only `^[a-z]{3,9}$`. The service checks local uniqueness, performs
        a Remnawave preflight, and also treats Remnawave `A019` as a duplicate if
        the upstream race is lost.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsernameRequest'
      responses:
        '200':
          description: Username reserved and onboarding advanced.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthState'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          $ref: '#/components/responses/UpstreamFailure'
  /api/v1/onboarding/agreement:
    post:
      tags: [Onboarding]
      operationId: acceptSharingAgreement
      summary: Accept the no-sharing agreement and provision Remnawave
      description: |
        Records acceptance and ensures an ACTIVE Remnawave user with a 2099
        upstream expiry, zero traffic, NO_RESET, and no squads. Recoverable partial
        failures remain resumable and are reconciled by username and Telegram ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgreementRequest'
      responses:
        '200':
          description: Agreement accepted and onboarding complete.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthState'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          $ref: '#/components/responses/UpstreamFailure'
  /api/v1/dashboard:
    get:
      tags: [Dashboard]
      operationId: getDashboard
      summary: Get the current home-page projection
      description: Returns balance, active and queued purchases, subscription data, and cached Remnawave statistics. `statisticsStale` is true when cached statistics were used after an upstream failure.
      responses:
        '200':
          description: Dashboard projection.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/catalog:
    get:
      tags: [Catalog]
      operationId: getCatalog
      summary: List purchasable combos and add-on squads
      responses:
        '200':
          description: Current visible catalog. The server will recalculate all prices at purchase time.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Catalog'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/purchases:
    get:
      tags: [Catalog]
      operationId: listPurchases
      summary: List the current user's purchase history
      responses:
        '200':
          description: Purchases newest first.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Purchase'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Catalog]
      operationId: createPurchase
      summary: Purchase or renew a combo with optional squads
      description: |
        Prices are loaded from the current catalog and debited atomically from the
        TXB ledger. First purchases activate immediately, same-plan renewals extend
        from the later of now or current expiry, and plan/add-on changes queue for
        the current term boundary. No proration is performed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseRequest'
      responses:
        '201':
          description: Purchase was created and its Remnawave operation was queued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Purchase'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Insufficient balance, outstanding debt, stale catalog choice, or incompatible queued change.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/subscription/revoke:
    post:
      tags: [Dashboard]
      operationId: revokeSubscription
      summary: Rotate the current Remnawave subscription URL
      description: Calls Remnawave revoke with `revokeOnlyPasswords:false`. The old URL must be treated as invalid after success.
      responses:
        '200':
          description: Rotated subscription reference.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionRotation'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          $ref: '#/components/responses/UpstreamFailure'
  /api/v1/balance:
    get:
      tags: [Billing]
      operationId: getBalance
      summary: Get current TXB balance
      responses:
        '200':
          description: Current ledger-derived balance and enabled-provider availability.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/ledger:
    get:
      tags: [Billing]
      operationId: listLedgerEntries
      summary: List the current user's immutable TXB ledger
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Ledger entries newest first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LedgerPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/payments/orders:
    post:
      tags: [Billing]
      operationId: createPaymentOrder
      summary: Create a balance top-up order
      description: |
        The requested TXB minor units are converted using the current fixed-decimal
        dashboard rate. Provider payable amounts are rounded upward to provider
        precision. A successful authoritative callback credits exactly `txbMinor`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentOrderRequest'
      responses:
        '201':
          description: Durable payment order, QR payload, and provider navigation target.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentOrder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          $ref: '#/components/responses/UpstreamFailure'
  /api/v1/payments/orders/{id}:
    get:
      tags: [Billing]
      operationId: getPaymentOrder
      summary: Poll durable payment state
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      responses:
        '200':
          description: Current authoritative order state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentOrder'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/webhooks/telegram:
    post:
      tags: [Webhooks]
      operationId: receiveTelegramWebhook
      summary: Receive Telegram bot updates
      description: Validates the configured Telegram webhook secret header. Relevant join requests, chat-member updates, pre-checkout queries, and successful payments are processed idempotently.
      security: []
      parameters:
        - name: X-Telegram-Bot-Api-Secret-Token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelegramUpdate'
      responses:
        '204':
          description: Update accepted or already processed.
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/webhooks/ezpay:
    get:
      tags: [Webhooks]
      operationId: receiveEZPayWebhook
      summary: Receive the signed EZPay asynchronous notification
      description: Only a valid signature, `TRADE_SUCCESS`, known order/provider trade, and exact payable amount can credit TXB. Replayed notifications are acknowledged without a second credit.
      security: []
      parameters:
        - name: pid
          in: query
          required: true
          schema: { type: string }
        - name: trade_no
          in: query
          required: true
          schema: { type: string }
        - name: out_trade_no
          in: query
          required: true
          schema: { type: string }
        - name: type
          in: query
          required: true
          schema: { type: string }
        - name: name
          in: query
          required: false
          schema: { type: string }
        - name: money
          in: query
          required: true
          schema:
            type: string
            pattern: '^\d+(\.\d{1,2})?$'
        - name: trade_status
          in: query
          required: true
          schema:
            type: string
            enum: [TRADE_SUCCESS]
        - name: sign
          in: query
          required: true
          schema: { type: string }
        - name: sign_type
          in: query
          required: true
          schema:
            type: string
            enum: [MD5]
      responses:
        '200':
          description: Plain-text `success` for a valid notification, including an idempotent replay.
          content:
            text/plain:
              schema:
                type: string
                const: success
        '400':
          description: Signature or notification validation failed.
          content:
            text/plain:
              schema:
                type: string
                const: fail
        '409':
          description: The signed payment did not match durable order state.
          content:
            text/plain:
              schema:
                type: string
                const: fail
  /api/v1/webhooks/bepusdt:
    post:
      tags: [Webhooks]
      operationId: receiveBEPusdtWebhook
      summary: Receive a signed BEPusdt transaction notification
      description: Status `2` is the only crediting state. Signature, local order, provider trade ID, requested USD amount, actual crypto amount, and recipient are checked before the idempotent ledger transaction.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BEPusdtNotification'
      responses:
        '200':
          description: Configured plain-text acknowledgement; defaults to `ok`.
          content:
            text/plain:
              schema:
                type: string
                example: ok
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/payments/return/{provider}:
    get:
      tags: [Billing]
      operationId: returnFromPaymentProvider
      summary: Navigate back to the Mini App after provider checkout
      description: This route never credits balance. It redirects to the client order screen, which polls durable webhook-confirmed state.
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [ezpay, bepusdt]
      responses:
        '303':
          description: Redirect to the same-origin payment result route.
          headers:
            Location:
              required: true
              schema:
                type: string
  /api/v1/payments/return/{provider}/{orderID}:
    get:
      tags: [Billing]
      operationId: returnFromPaymentProviderWithOrder
      summary: Navigate back to a specific Mini App payment order
      description: Compatibility return route for providers that preserve path parameters more reliably than query parameters. It never credits balance.
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [ezpay, bepusdt]
        - name: orderID
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '303':
          description: Redirect to the same-origin payment result route.
          headers:
            Location:
              required: true
              schema:
                type: string
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: txc_session
      description: Opaque seven-day session created only by validated Telegram Mini App init data.
  parameters:
    ResourceID:
      name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
...
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor returned by the preceding page.
      schema:
        type: string
    PageSize:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Missing, invalid, or expired authentication or provider signature.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: The authenticated identity cannot perform this operation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Resource not found or not visible to the caller.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: Request conflicts with current persisted state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    UpstreamFailure:
      description: A required provider could not complete the synchronous portion of the request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    ID:
      type: string
      minLength: 1
      examples: [01JZX6BKW16M0CNQ5Q74K8DN7J]
...
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]*$'
          examples: [INSUFFICIENT_BALANCE]
        message:
          type: string
        details:
          type: object
          additionalProperties:
            type: string
        requestId:
          type: string
          minLength: 1
    TelegramAuthRequest:
      type: object
      additionalProperties: false
      required: [initData]
      properties:
        initData:
          type: string
          minLength: 1
          description: Raw `Telegram.WebApp.initData` query string.
...
    AuthState:
      type: object
      additionalProperties: false
      required: [authenticated, user]
      properties:
        authenticated:
          type: boolean
          const: true
        user:
          $ref: '#/components/schemas/User'
...
    OnboardingInvites:
      type: object
      additionalProperties: false
      required: [group, channel]
      properties:
        group:
          $ref: '#/components/schemas/JoinInvite'
        channel:
          $ref: '#/components/schemas/JoinInvite'
    MembershipState:
      type: object
      additionalProperties: false
      required: [groupJoined, channelJoined, complete, user]
      properties:
        groupJoined:
          type: boolean
        channelJoined:
          type: boolean
        complete:
          type: boolean
        user:
          $ref: '#/components/schemas/User'
    UsernameRequest:
      type: object
      additionalProperties: false
      required: [username]
      properties:
        username:
          type: string
          pattern: '^[a-z]{3,9}$'
    AgreementRequest:
      type: object
      additionalProperties: false
      required: [accepted]
      properties:
        accepted:
          type: boolean
          const: true
...
    BalanceResponse:
      type: object
      additionalProperties: false
      required: [balance, paymentMethods]
      properties:
        balance:
          $ref: '#/components/schemas/Money'
        paymentMethods:
          type: array
          items:
            $ref: '#/components/schemas/PaymentMethod'
...
    Catalog:
      type: object
      additionalProperties: false
      required: [combos, addons]
      properties:
        combos:
          type: array
          items:
            $ref: '#/components/schemas/Combo'
        addons:
          type: array
          items:
            $ref: '#/components/schemas/SquadProduct'
...
    PurchaseRequest:
      type: object
      additionalProperties: false
      required: [comboId, addonSquadProductIds]
      properties:
        comboId:
          $ref: '#/components/schemas/ID'
        addonSquadProductIds:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ID'
...
    Purchase:
      type: object
      additionalProperties: false
      required:
        - id
        - comboId
        - comboName
        - price
        - validFrom
        - validUntil
        - status
        - trafficLimitBytes
        - resetStrategy
        - squadUuids
        - createdAt
        - updatedAt
      properties:
        id:
          $ref: '#/components/schemas/ID'
        comboId:
          $ref: '#/components/schemas/ID'
        comboName:
          type: string
        price:
          $ref: '#/components/schemas/Money'
        validFrom:
          $ref: '#/components/schemas/Timestamp'
        validUntil:
          $ref: '#/components/schemas/Timestamp'
        status:
          $ref: '#/components/schemas/EntitlementStatus'
        trafficLimitBytes:
          type: string
          pattern: '^\d+$'
        resetStrategy:
          $ref: '#/components/schemas/ResetStrategy'
        squadUuids:
          type: array
          items:
            type: string
            format: uuid
        createdAt:
          $ref: '#/components/schemas/Timestamp'
        updatedAt:
          $ref: '#/components/schemas/Timestamp'
    SubscriptionRotation:
      type: object
      additionalProperties: false
      required: [subscriptionUrl]
      description: The URL is a bearer credential and must not be logged, cached outside the authenticated UI, or exposed through admin list endpoints.
      properties:
        subscriptionUrl:
          type: string
          format: uri
...
    Dashboard:
      type: object
      additionalProperties: false
      required:
        - user
        - balance
        - activePurchase
        - queuedPurchase
        - statistics
        - subscriptionUrl
        - statisticsStale
        - statisticsWarning
        - fetchedAt
      properties:
        user:
          $ref: '#/components/schemas/User'
        balance:
          $ref: '#/components/schemas/Money'
        activePurchase:
          oneOf:
            - $ref: '#/components/schemas/Purchase'
            - type: 'null'
        queuedPurchase:
          oneOf:
            - $ref: '#/components/schemas/Purchase'
            - type: 'null'
        statistics:
          oneOf:
            - $ref: '#/components/schemas/Statistics'
            - type: 'null'
        subscriptionUrl:
          type: [string, 'null']
          format: uri
        statisticsStale:
          type: boolean
        statisticsWarning:
          type: string
        fetchedAt:
          $ref: '#/components/schemas/Timestamp'
...
    LedgerPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/LedgerEntry'
        page:
          $ref: '#/components/schemas/PageInfo'
...
    PaymentOrderRequest:
      type: object
      additionalProperties: false
      required: [provider, txbMinor]
      properties:
        provider:
          $ref: '#/components/schemas/PaymentProvider'
        txbMinor:
          type: string
          pattern: '^[1-9]\d*$'
...
    BEPusdtNotification:
      type: object
      additionalProperties: true
      required: [trade_id, order_id, amount, actual_amount, token, status, signature]
      properties:
        trade_id:
          type: string
        order_id:
          type: string
        amount:
          $ref: '#/components/schemas/FlexibleDecimal'
        actual_amount:
          $ref: '#/components/schemas/FlexibleDecimal'
        token:
          type: string
        block_transaction_id:
          type: string
        status:
          type: integer
          enum: [1, 2, 3]
        signature:
          type: string
        created_at:
          type: string
        expired_at:
          type: string
...